BUG: Replace itk::Math::Round template argument `float` with `int`
According to https://github.com/InsightSoftwareConsortium/ITK/blob/66ab5b14c476d057ed46a4daba05b710681160e6/Modules/Core/Common/include/itkMath.h#L127 the `TReturn` template argument of `Math::Round` must be an integer type. In practice, it will internally call `Detail::RoundHalfIntegerToEven_32` when `TReturn` is a 32-bit `float`. Which returns an `int32_t`, which is equivalent to `int`.

+++ Core/elxProgressCommand.cxx
@@ -167,7 +167,7 @@ void
 ProgressCommand::PrintProgress(const float progress) const
 {
   /** Print the progress to the screen. */
-  const int progressInt = itk::Math::Round<float>(100 * progress);
+  const int progressInt = itk::Math::Round<int>(100 * progress);
 
   // Pass the entire message at once, rather than having multiple `<<` insertions.
   const std::string message = '\r' + m_StartString + std::to_string(progressInt) + m_EndString;
